[SPARK-49323][CONNECT] Move MockObserver from Spark Connect Server's test folder to the Server's main folder
### What changes were proposed in this pull request?

Move class `MockObserver` from the Spark Connect Server's test folder to the Server's main folder, so that this class can be shaded as well, from `io.grpc` to `org.sparkproject.connect.grpc`.

https://github.com/apache/spark/blob/02f4e5436c9b41b1e6dfecc2ce69b9613ef46428/project/SparkBuild.scala#L775

### Why are the changes needed?

The `MockObserver` class is a mock of the `StreamObserver` class from the gRPC library, used for testing the process function of [SparkConnectPlanner](url).

Currently, the [[`MockObserver`]] class is located in the test folder, which is not shaded by sbt-assembly, unlike the main folder.

This results in a compilation error when open-source libraries call the `transform(spark: SparkSession, command: proto.Command)` function in `SparkConnectPlannerTestUtils`, which in turn calls `SparkConnectPlanner(executeHolder).process(command, new MockObserver())`.

To resolve this, the `MockObserver` class should be moved to the main folder to be shaded as well.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing UTs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #47816 from longvu-db/move-mock-observer.

Authored-by: Thang Long VU <long.vu@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>