[SPARK-49336][CONNECT] Limit the nesting level when truncating a protobuf message
### What changes were proposed in this pull request?

Abbreviating a protobuf message is ineffective if the message is very deeply nested, causing a serious performance problem when converting the message into a string. This PR adds a new parameter, maxLevel, to ProtoUtils.abbreviate which limits the maximum nesting level to abbreviate and truncate any data that is below the level. As a result, ProtoUtils.abbreviate may return a partial protobuf message in case the message is deeply nested and a small value of maxLevel is supplied.

### Why are the changes needed?

Described above.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

"truncate nested message" added to AbbreviateSuite.scala.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #47824 from changgyoopark-db/SC-174553.

Authored-by: Changgyoo Park <changgyoo.park@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>