[SPARK-49318][SQL] Preempt low priority error on LCA until end of check analysis to improve error experience
### What changes were proposed in this pull request?
This PR allows high-priority analysis errors to be thrown first when a low-priority error on LCA (lateral column alias) occurs in the same query.
It preempts and stores the LCA error during check analysis. If another error happens, it will be thrown first; If not, the LCA error will be thrown at the end of the analysis.

### Why are the changes needed?
This helps improve the user experience on errors, to provide higher priority information and the root cause to users.
The test in the PR is one case that UNRESOLVED_COLUMN should be thrown, instead of the LCA internal error, which was thrown without the change because the segment sits in the CTE definition that is traversed first in CheckAnalysis.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
A new test case.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #47813 from anchovYu/lca-error-priority.

Authored-by: Xinyi Yu <xinyi.yu@databricks.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>