[SPARK-49113] Dont assert on translation bugs - silently swallow the exception
### What changes were proposed in this pull request?

### Why are the changes needed?

In this PR I propose that we do not assert and fail queries when `V2ExpressionBuilder` returns class that is not predicate. Instead I suggest that we silently swallow the exception, add log, and continue execution of query.

`V2ExpressionBuilder` has no compiler gurantees that `Predicate` will be returned. With this in mind, its very easy to add regressions that will suddenly start failing queries, without a workaround for users, other than code change.

In this PR, I suggest we add a flag, which when enabled, will absorb this bugs in `V2ExpressionBuilder`.

It is as if we are saying:
 - If we are trying to translate catalyst expression to `V2Expression` and are requesting `Predicate`, if predicate is not returned act as if we didn't know how to translate expression at all.

We return None, and rest of the system handles it as if we didn't know to translate to `V2Expression`

More details in JIRA ticket

### Does this PR introduce _any_ user-facing change?

Queries that failed because of pushdown failure, will now skip pushing down offending predicate, rather than failing the whole query.

### How was this patch tested?

Unit tests.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #47611 from milastdbx/dev/milast/removeAssert.

Authored-by: milastdbx <milan.stefanovic@databricks.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>