[SPARK-49345][CONNECT] Make sure using the current running Spark Session
### What changes were proposed in this pull request?

This PR proposes to make sure using the current running Spark Session.

### Why are the changes needed?

If users use `SparkSession.active()` or `SQLConf.get`, then it might pick up wrong configurations from a different Spark session. The current codebase is not affected so this is more for safety in the future.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Not added because the current codebase is not affected. This is just more for completeness.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #47838 from HyukjinKwon/set-active-session.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>