[SPARK-49306][SQL] Create new SQL functions 'zeroifnull' and 'nullifzero'
### What changes were proposed in this pull request?

This PR creates new SQL functions `zeroifnull` and `nullifzero`.

These are equivalent to `nvl(input, 0)` and `if(input = 0, null, input)`, respectively. The analyzer performs these replacements inline in this PR to take advantage of existing query optimizations for `nvl` and `if` expressions.

For example:

```
> select zeroifnull(null);
 0
> select zeroifnull(2);
 2

> select nullifzero(0);
 null
> select nullifzero(2);
 2
```

### Why are the changes needed?

We are adding these functions for compatibility with other systems which also have these SQL functions.

### Does this PR introduce _any_ user-facing change?

Yes, see above.

### How was this patch tested?

This PR adds unit test coverage.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #47817 from dtenedor/zeroifnull.

Authored-by: Daniel Tenedorio <daniel.tenedorio@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>