[SPARK-49385][K8S] Fix `getReusablePVCs` to use `podCreationTimeout` instead of `podAllocationDelay`
### What changes were proposed in this pull request?

This PR aims to use `podCreationTimeout` instead of `podAllocationDelay` when `getReusablePVCs` excludes the newly created PVCs of previous batches.

### Why are the changes needed?

K8s control plane pod creation can be delayed due to the unknown reasons. So, `podAllocationDelay (default: 1s)` is insufficient to say that the previous allocation batch's pods are created with their PVCs. We had better wait until `podCreationTimeout`.

### Does this PR introduce _any_ user-facing change?

This affects only the initial set of executors because the baseline is PVC's `getCreationTimestamp`. So, this fixes only a buggy situation where a PVC is shared by two executors due to the long pending executor pod.

### How was this patch tested?

Pass the CIs with newly updated test cases.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #47867 from dongjoon-hyun/SPARK-49385.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>