[SPARK-49363][SS][TESTS] Add unit tests for potential RocksDB state store SST file mismatch
### What changes were proposed in this pull request?
Add unit test to for RocksDB state store snapshot checkpointing for changelog. We intentionally add the same content in each batch, so that it is likely that SST files generated are all of the same size. We have some randomness on loading the existing version or move to the next, and whether maintenance task is executed. All three tests would fail for previous versions but not in master.

### Why are the changes needed?
Recently we discovered some RocksDB state store file version ID mismatch issues. Although it happens to have been fixed by other change, we don't have test coverage for it. Add unit tests for them.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Run the tests

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #47850 from siying/cp_test.

Authored-by: Siying Dong <siying.dong@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>