[SPARK-49492][CONNECT] Reattach attempted on inactive ExecutionHolder
### What changes were proposed in this pull request?

Check the status of the ExecutionHolder before reattaching it to the client.

### Why are the changes needed?

An ExecutionHolder may fail to spawn an ExecuteThreadRunner if an exception is thrown before doing so. In that case, reattaching the ExecutionHolder will succeed but no progress will be made.
-> *As a result, the job is hanging.*

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Add a test case to ReattachableExecuteSuite.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #47955 from changgyoopark-db/SPARK-49492.

Authored-by: Changgyoo Park <changgyoo.park@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>