[SPARK-49591][SQL] Add Logical Type column to variant readme
### What changes were proposed in this pull request?

Add a concept of logical type to the variant README.md, distinct from the physical encoding of a value. In particular, decimal and integer values are considered to be members of a single "Exact Numeric" type.

### Why are the changes needed?

This is intended to describe and justify the existing Spark behaviour for Variant (e.g. stripping trailing zeros for decimal to string casts), not change it. (Although the SchemaOfVariant expression does not strictly follow this right now for numeric types, and should be updated to match it.) The motivation for introducing a single numeric type that encompasses integer and decimal values is to allow more flexibility in storage (particularly once shredding is introduced), and provide a simpler user surface, since there is not much benefit to distinguishing numeric precision/scale on a per-value basis.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

It is a documentation change.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48064 from cashmand/cashmand/SPARK-49591.

Authored-by: cashmand <david.cashman@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>