[SPARK-49392][SQL] Catch errors when failing to write to external data source
### What changes were proposed in this pull request?
Catch various exceptions thrown by the data source API, when failing to write to a data source, and rethrow `externalDataSourceException` to provide a more friendly error message for the user.

### Why are the changes needed?
To catch non-fatal exceptions when failing to save the results of a query into an external data source (for example: `com.crealytics.spark.excel`).

### Does this PR introduce _any_ user-facing change?
Yes, error messages when failing to write to an external data source should now be more user-friendly.

### How was this patch tested?
New test in `ExplainSuite`.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #47873 from uros-db/external-data-source.

Authored-by: Uros Bojanic <uros.bojanic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>