[SPARK-49585][CONNECT] Replace executions map in SessionHolder with operationID set
### What changes were proposed in this pull request?

SessionHolder has no reason to store ExecuteHolder directly as SparkConnectExecutionManager has a global map of ExecuteHolder. This PR replaces the map in SessionHolder with a set of operation IDs which is only used when interrupting executions within the session.

### Why are the changes needed?

Save memory, and simplify the code by making SparkConnectExecutionManager the single source of ExecuteHolders.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48071 from changgyoopark-db/SPARK-49585.

Authored-by: Changgyoo Park <changgyoo.park@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>