[SPARK-49369][CONNECT][SQL] Add implicit Column conversions
### What changes were proposed in this pull request?
This introduces an implicit conversion for the Column companion object that allows a user/developer to create a Column from a catalyst Expression (for Classic) or a proto Expression (Builder) (for Connect). This mostly recreates they had before we refactored the Column API. This comes at the price of adding the an import.

### Why are the changes needed?
Improved upgrade experience for Developers and User who create their own Column's from expressions.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
I added it to a couple of places in the code and it works.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48020 from hvanhovell/SPARK-49369.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>