[SPARK-49861][INFRA] Add `Python 3.13` to Infra docker image
### What changes were proposed in this pull request?

This PR aims to add `Python 3.13` to Infra docker image.

Note that SPARK-49862 `Add BASIC_PIP_PKGS and CONNECT_PIP_PKGS to Python 3.13` will track the Python package readiness.

### Why are the changes needed?

`Python 3.13` release is scheduled on next Monday, 2024-10-07.
- https://peps.python.org/pep-0719/

This is a part of `Python 3.13 Support` preparation for Apache Spark 4.0.0 on 2025 February.

### Does this PR introduce _any_ user-facing change?

No, this is an infra only change.

### How was this patch tested?

Pass the CIs and verify manually like the following.
```
$ docker run -it --rm ghcr.io/dongjoon-hyun/apache-spark-ci-image:master-11152626644 python3.13 --version
Python 3.13.0rc3
```

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48330 from dongjoon-hyun/SPARK-49861.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>