[SPARK-49867][SQL] Improve the error message when index is out of bounds when calling GetColumnByOrdinal
### What changes were proposed in this pull request?
Provide more user facing error when ordinal of the column is out of bounds.

### Why are the changes needed?
There's an issue where `GetColumnByOrdinal` can return attributes which length doesn't match ordinal which is used as a parameter. When that happens we throw an assertion error which is not user friendly. Because of that we introduced new QueryExecutionError in order to make it more user facing.

### Does this PR introduce _any_ user-facing change?
Yes, error message is more user friendly.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #48339 from mihailoale-db/mihailoale-db/fixgetcolumnordinaloutofboundserrormessage.

Authored-by: Mihailo Aleksic <mihailo.aleksic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>