[SPARK-49898][CORE] Fix documentation and default for event log task metrics accumulator logging flag from SPARK-42204
### What changes were proposed in this pull request?

This PR corrects an unintentional default behavior change from https://github.com/apache/spark/pull/39763

That PR introduced a new configuration, `spark.eventLog.includeTaskMetricsAccumulators`, to provide an ability for users to disable the redundant logging of task metrics information via the Accumulables field in the Spark event log task end logs.

I made a mistake in updating that PR description and code from the original version: the description says that the intent is to not change out of the box behavior, but the actual flag default was the opposite.

This new PR corrects both the flag default and the flag description to reflect the original intent of not changing default behavior.

### Why are the changes needed?

Roll back an unintentional behavior change.

### Does this PR introduce _any_ user-facing change?

Yes, it rolls back an unintentional default behavior change.

### How was this patch tested?

Existing unit tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #48372 from JoshRosen/fix-event-log-accumulable-defaults.

Authored-by: Josh Rosen <joshrosen@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>