[SPARK-49927][SS] pyspark.sql.tests.streaming.test_streaming_listener to wait longer
### What changes were proposed in this pull request?
In test pyspark.sql.tests.streaming.test_streaming_listener, instead of waiting for fixed 10 seconds, we wait for progress made.

### Why are the changes needed?
In some environment, the test fails with progress_event appears to be None. Likely the wait time is not sufficient.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Run the patch and make sure it still passes

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #48414 from siying/python_test.

Lead-authored-by: Siying Dong <siying.dong@databricks.com>
Co-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>