[SPARK-49915][SQL] Handle zeros and ones in ReorderAssociativeOperator
### What changes were proposed in this pull request?

For additions, we omit the `Add` operation if the foldable ones finally result in 0, e.g. `-3 + a + 3` is simplified to `a` instead of `a + 0`.
For multiplication,
- we simplify the expression to `Literal(0, dt)` if the foldable ones finally result in 0 && the expression itself isn't nullable
- we omit the `Multiply` operation if the foldable ones finally result in 1

### Why are the changes needed?

Improve the simplicity of expression evaluation and the opportunities for predicates to be pushed down to data sources

### Does this PR introduce _any_ user-facing change?

no, the result shall be identical

### How was this patch tested?

new tests

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #48395 from yaooqinn/SPARK-49915.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Kent Yao <yao@apache.org>